requirements.txt
pytest
pytest-html
pytest-xdist
PyYAML
